refactor(Collapsible): rename useCollapsible to useXDSCollapsible
All public hooks use the `useXDS` prefix per XDS naming conventions.

Renamed:
- useCollapsible → useXDSCollapsible
- UseCollapsibleOptions → UseXDSCollapsibleOptions
- UseCollapsibleReturn → UseXDSCollapsibleReturn
- useCollapsible.ts → useXDSCollapsible.ts
